refactor: unify UploadService implementations into a single service
