+++ apps/meteor/app/api/server/v1/users.ts
@@ -1,6 +1,6 @@
 import { readFile } from 'fs/promises';
 
-import { MeteorError, Team, api, Calendar } from '@rocket.chat/core-services';
+import { MeteorError, Team, api, Calendar, Upload } from '@rocket.chat/core-services';
 import { type IExportOperation, type ILoginToken, type IPersonalAccessToken, type IUser, type UserStatus } from '@rocket.chat/core-typings';
 import { Users, Subscriptions, Sessions } from '@rocket.chat/models';
 import {
@@ -77,7 +77,6 @@ import { API } from '../api';
 import { getPaginationItems } from '../helpers/getPaginationItems';
 import { getUserFromParams } from '../helpers/getUserFromParams';
 import { isUserFromParams } from '../helpers/isUserFromParams';
-import { UploadService } from '../lib/UploadService';
 import { isValidQuery } from '../lib/isValidQuery';
 import { findPaginatedUsersByStatus, findUsersToAutocomplete, getInclusiveFields, getNonEmptyFields, getNonEmptyQuery } from '../lib/users';
 
@@ -272,7 +271,7 @@ API.v1.addRoute(
 				return API.v1.success();
 			}
 
-			const { file, fields } = await UploadService.parse(this.request, {
+			const { file, fields } = await Upload.parseUpload(this.request, {
 				field: 'image',
 				maxSize: settings.get<number>('FileUpload_MaxFileSize'),
 			});
